% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualMLE-package.R
\docType{package}
\name{visualMLE-package}
\alias{visualMLE-package}
\title{Visual Maximum Likelihood Estimation}
\description{
Visual Maximum Likelihood Estimation
}
\author{
\email{jason@bryer.org}
}
\keyword{MLE}
\keyword{estimation}
\keyword{likelihood}
\keyword{maximum}
